requirements.txt
fastapi
uvicorn[standard]
python-dotenv
couchbase
gunicorn
httpx
pydantic
pydantic
prometheus-fastapi-instrumentator